Move Gitlab merge request to gitlab/mr
